+++ include/bitprim/nodecint/chain/compact_block.h
@@ -39,13 +39,13 @@ BITPRIM_EXPORT
 int /*bool*/ compact_block_is_valid(compact_block_t block);
 
 BITPRIM_EXPORT
-size_t compact_block_serialized_size(compact_block_t block, uint32_t version);
+uint64_t /*size_t*/ compact_block_serialized_size(compact_block_t block, uint32_t version);
 
 BITPRIM_EXPORT
-size_t compact_block_transaction_count(compact_block_t block);
+uint64_t /*size_t*/ compact_block_transaction_count(compact_block_t block);
 
 BITPRIM_EXPORT
-transaction_t compact_block_transaction_nth(compact_block_t block, size_t n);
+transaction_t compact_block_transaction_nth(compact_block_t block, uint64_t /*size_t*/ n);
 
 BITPRIM_EXPORT
 uint64_t compact_block_nonce(compact_block_t block);
